moved initial state from store to reducer, refactored code and added condition to server to guard against user entering empty symbol
