+++ stock-tracker/server.js
@@ -27,6 +27,10 @@ io.on("connection", socket => {
     getSearchInputAndFilter(socket, searchInput, stockCompanies);
   });
   socket.on("symbol", (stockSymbol, chartTime) => {
+    if (stockSymbol === "") {
+      Object.values(intervals).forEach(clearInterval);
+      return;
+    }
     console.info("Stock Symbol Entered: " + stockSymbol);
     Object.values(intervals).forEach(clearInterval);
     intervals.stock = callNowAndInterval(
